+++ includes/session_init.php
@@ -1,14 +1,16 @@
 <?php
 
+require_once __DIR__ . '/security_headers.php';
+
 if (!isset($_SESSION)) {
     // HTTP Only cookies
     ini_set("session.cookie_httponly", true);
-    
+
     if ($config_https_only) {
         // Tell client to only send cookie(s) over HTTPS
         ini_set("session.cookie_secure", true);
     }
-    
+
     session_start();
 
 }